Make Spring Security play with @TestSecurity
Follows up on: #10487
Resolves: #10490
