+++ integration-tests/spring-web/src/main/resources/application.properties
@@ -3,3 +3,7 @@ quarkus.security.users.file.users=test-users.properties
 quarkus.security.users.file.roles=test-roles.properties
 quarkus.security.users.file.plain-text=true
 %test.quarkus.http.test-port=0
+
+# we add this because we also use @TestSecurity which means that basic auth is disabled by default because @TestSecurity
+# contributes TestHttpAuthenticationMechanism
+%test.quarkus.http.auth.basic=true
