+++ integration-tests/spring-web/src/test/java/io/quarkus/it/spring/web/SecurityIT.java
@@ -1,0 +1,7 @@
+package io.quarkus.it.spring.web;
+
+import io.quarkus.test.junit.NativeImageTest;
+
+@NativeImageTest
+public class SecurityIT extends SecurityTest {
+}
